requirements.txt
pandas
plotly
requests
streamlit
numpy
Pillow
tensorflow